requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
python-dotenv==1.0.0
pytest
pytest-flask
pytest-dotenv
gunicorn==21.2.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9